move find_resource to state.rs
